ver2/requirements.txt
# Core
numpy
pydantic>=2
python-dotenv

# Data loading
requests
beautifulsoup4
lxml
trafilatura
pypdf

# Embeddings / Vector index
sentence-transformers
faiss-cpu

# LLM
torch
transformers>=4.41
accelerate

# Web UI
gradio

# Utils
tqdm
